Move NewClefTransactor to auth_clef.go
